Move 32-bit integer into primitives
