+++ library/Rattletrap/AttributeValue/FlaggedInt.hs
@@ -1,6 +1,6 @@
 module Rattletrap.AttributeValue.FlaggedInt where
 
-import Rattletrap.Int32
+import Rattletrap.Primitive.Int32
 
 import qualified Data.Binary.Bits.Get as BinaryBit
 import qualified Data.Binary.Bits.Put as BinaryBit
