+++ library/Rattletrap/AttributeValue/WeldedInfo.hs
@@ -1,7 +1,7 @@
 module Rattletrap.AttributeValue.WeldedInfo where
 
 import Rattletrap.Primitive.Float32
-import Rattletrap.Int32
+import Rattletrap.Primitive.Int32
 import Rattletrap.Primitive.Int8Vector
 import Rattletrap.Vector
