+++ library/Rattletrap/Header.hs
@@ -1,7 +1,7 @@
 module Rattletrap.Header where
 
 import Rattletrap.Primitive.Dictionary
-import Rattletrap.Int32
+import Rattletrap.Primitive.Int32
 import Rattletrap.Property
 import Rattletrap.PropertyValue
 import Rattletrap.Text
