+++ library/Rattletrap/PropertyValue.hs
@@ -2,7 +2,7 @@ module Rattletrap.PropertyValue where
 
 import Rattletrap.Primitive.Dictionary
 import Rattletrap.Primitive.Float32
-import Rattletrap.Int32
+import Rattletrap.Primitive.Int32
 import Rattletrap.List
 import Rattletrap.Text
 import Rattletrap.Word64
